requirements.txt
asgiref==3.7.2
attrs==23.1.0
cffi==1.15.1
cryptography==41.0.1
Django==4.2.1
django-cors-headers==4.0.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-resources==5.12.0
jsonschema==4.17.3
pkgutil_resolve_name==1.3.10
postgres==4.0
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pycparser==2.21
pyproject-toml==0.0.10
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.6.3
zipp==3.15.0
